修改跟进定位类型

+++ app/src/main/java/cn/atsoft/dasheng/app/entity/BusinessTrack.java
@@ -104,13 +104,13 @@ public class BusinessTrack implements Serializable {
      * 经度
      */
     @TableField("longitude")
-    private BigDecimal longitude;
+    private Double longitude;
 
     /**
      * 纬度
      */
     @TableField("latitude")
-    private BigDecimal latitude;
+    private Double latitude;
 
     /**
      * 负责人
@@ -239,19 +239,19 @@ public class BusinessTrack implements Serializable {
         this.image = image;
     }
 
-    public BigDecimal getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(BigDecimal longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 
-    public BigDecimal getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(BigDecimal latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
